feat(all) little modifications

+++ src/main/java/com/kenjerdb/core/dao/FileDatabaseReader.java
@@ -4,7 +4,7 @@ import java.io.File;
 
 public interface FileDatabaseReader {
 
-    String recordByIndex(File readable, String delimiter, int index);
+    String recordByIndex(File readable, String delimiter, Long index);
 
     String readAll(File readable);
 }
